правки по замечаниям
