+++ test/com/shishkin/tasktracker/service/FileBackedTaskManagerTest.java
@@ -22,7 +22,6 @@ public class FileBackedTaskManagerTest extends TaskManagerTest<FileBackedTaskMan
 
     private File file;
 
-    @Override
     @BeforeEach
     void setUp() {
         file = new File("./resources/tasks_test.csv");
